+++ CHANGELOG.md
@@ -20,6 +20,7 @@
   * Fixed bracket operator not working for steps within the pipeline ([#42](https://github.com/KIT-IAI/pyWATTS/issues/42))
   * Fixed dict objects could not be passed to pipeline ([#43](https://github.com/KIT-IAI/pyWATTS/issues/43))
   * Fixed old parameter in sample_module  ([#67](https://github.com/KIT-IAI/pyWATTS/issues/67))
+  * Fixed array creation in trend extraciont ([#69](https://github.com/KIT-IAI/pyWATTS/issues/69))
 
 ## [0.0.1] - 2021-MM-DD